Conteneurisation/requirements_api.txt
fastapi
uvicorn
pydantic
pandas
mlflow
psycopg2-binary

